requirements.txt
# Core ML
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
pillow==10.2.0
scikit-learn==1.4.0

# Experiment tracking
mlflow==2.10.2

# API
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1

# Monitoring
prometheus-client==0.19.0

# Data versioning
dvc==3.42.0

# Utilities
tqdm==4.66.2
pyyaml==6.0.1
python-dotenv==1.0.1
requests==2.31.0
